ARM: OMAP5: clock-data: Add alias node for timer's clock source
In commit c59b537 (ARM: OMAP2+: Simplify dmtimer clock aliases), new
clock aliases for dmtimers were added to simplify the code. These clock
aliases can also be used when configuring the system timers and allow us
to remove the current definitions, simplifying the code. The clockdata
generated for OMAP5 is not adapted for this change.

So removing a seperate alias for every timer's 32k source and adding
the unique timer 32k-src clk node as it is in other OMAP clk data.

Signed-off-by: Sricharan R <r.sricharan@ti.com>