PM / AVS: Introduce support for OMAP Voltage Controller(VC) with device tree nodes
Texas Instrument's OMAP SoC generations OMAP3-5 introduced an TI
custom hardware block to better facilitate and standardize integration
of Power Management ICs(PMIC) which communicate over I2C. A key
hardware block that communicates physically with the PMIC is Voltage
Controller(VC).

This specialized hardware block meant solely to support PMIC and is
customized to that requirement. Even though it functions as an I2C
controller, it is a write-only interface whose configurations are
custom to PMICs.

We have an existing implementation in mach-omap2 which has been
re-factored and moved out as a separate driver. This new driver is DT
only and the separation was meant to get a maintainable driver which
does not have to deal with legacy platform_data dependencies. The legacy
driver is retained to support non-DT boot and functionality will be
migrated to the DT-only version as we enable features.

Currently, this implementation considers the basic steps needed for
voltage scaling and exposing voltage controller as a device whose
child devices are voltage controller channel devices.

We may need to do additional timing configurations to enable Low power
mode voltage transitions, but this will be completed in a following
series. We may further need a few tweaks to hook the Adaptive Voltage
Scaling(AVS) implementation for OMAP which also uses the same voltage
controller to talk to PMICs.

This driver is meant to interface with voltage processor when the
voltage processor attempts devm_omap_vc_channel_get for the VC channel
corresponding to the voltage processor.

[grygorii.strashko@ti.com, taras@ti.com: co-developer]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Taras Kondratiuk <taras@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>