PM / AVS: Introduce support for OMAP Voltage Processor(VP) with device tree nodes
Texas Instrument's OMAP SoC generations OMAP3-5 introduced an TI
custom hardware block to better facilitate and standardize integration
of Power Management ICs(PMIC) which communicate over I2C. A key
hardware block that communicates with the PMIC is Voltage
Processor(VP), which in turn communicates with Voltage Controller(VC)
for the required voltage to be set.

Voltage Processor is an specialized hardware block meant to support
PMIC chips only using VC. VP also provides an interface for
SmartReflex Adaptive Voltage Scaling(AVS) hardware module to send
adjustment steps which is converted into voltage values and send
onwards by VP to VC. VP is also used to set the voltage of the PMIC by
commanding using "forceupdate".

We have an existing implementation in mach-omap2 which has been
re-factored and moved out as a separate driver. This new driver is DT
only and the separation was meant to get a maintainable driver which
does not have to deal with legacy platform_data dependencies. The legacy
driver is retained to support non-DT boot and functionality will be
migrated to the DT-only version as we enable features.

Currently, this implementation considers only the basic steps needed for
voltage scaling and exposing voltage processor which hooks on to Voltage
Controller driver and OMAP PMIC driver to provide an regulator interface
over OMAP PMIC driver.

We may need to do additional timing configurations to enable Low power
mode voltage transitions and to hook AVS implementation for OMAP which
also uses the same voltage controller to talk to PMICs. This needs to
be addressed in a later series.

This driver is meant to interface with OMAP PMIC driver when the
controller driver registers it's operations with OMAP PMIC driver and
associates with an voltage controller channel. This enables the full
communication path between the OMAP PMIC regulator to the external PMIC
hardware over OMAP Voltage Processor and OMAP Voltage Controller.

[grygorii.strashko@ti.com, taras@ti.com: co-developer]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Taras Kondratiuk <taras@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>