xenbus: delay xenbus frontend resume if xenstored is not running
If the xenbus frontend is located in a domain running xenstored, the device
resume is hanging because it is happening before the process resume. This
patch adds extra logic to the resume code to check if we are the domain
running xenstored and delay the resume if needed.

Signed-off-by: Aurelien Chartier <aurelien.chartier@citrix.com>

Changes in v2:
- Instead of bypassing the resume, process it in a workqueue
Changes in v3:
- Add a struct work in xenbus_device to avoid dynamic allocation
- Several small code fixes
Changes in v4:
- Use a dedicated workqueue
Changes in v5:
- Move create_workqueue error handling to xenbus_frontend_dev_resume
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>