regulator: twl-regulator: allow for setting voltage offset where allowed.
In TWL603[02] depending on the value of SMPS_OFFSET efuse register
the voltage range supported in standard mode can be either
between 0.6V - 1.3V or 0.7V - 1.4V.

However the same chip can be produced with differing internal firmware
which may or may not allow for offset to be reprogrammed on the device.
Handle this on the fly and provide device tree bindings for the same.

Signed-off-by: Nishanth Menon <nm@ti.com>