ARM: OMAP4+: pm_debug: provide more visibility into suspend failure
Traditionally on OMAP4+ systems, with complete data properly populated,
we should be able to track down which device prevented the power domain
from transitioning to low power state. This is rather trivial is the
powerdomain to clockdomain to list of clock status registers were easily
accessible. However, it is not. we have linkage from clk->clkdm, and
hwmod to clkdm, however going from clkdm to clk status involves a search
for matches :(.

Previously, we just get:
Powerdomain (XYZ) didn't enter target state 1

Now, we get:
Powerdomain (XYZ) didn't enter target state 1
	XYZ: s/w flags(HWSUP_SWSUP HWSUP_ENABLED(int) ) h/w control(enable_auto )
		module-X: functional

This simplifies our debug when hwmod is properly populated with every
device in the SoC and when status bits are properly identified. However,
the alternative to doing this is to replicate actual h/w data in some
other form which duplicates existing infrastructure for the same.

Signed-off-by: Nishanth Menon <nm@ti.com>