ARM: dts: DRA7: Fix peripheral interrupt numbers for crossbar inputs (squash me)
commit (ARM: DTS: DRA7: Replace peripheral interrupt numbers with crossbar
inputs) has a quiet a few errors considering that it was not rebased to
latest linux-next state.

There are new peripherals added whose cross bar routing is missing as
well as ones such as DMA which was plain wrong.

Since the original commit from Sricharan's tree is not merged, should be
squashable.

Signed-off-by: Nishanth Menon <nm@ti.com>