irqchip: crossbar: introduce option for a 'irqs-safe-map'
Since crossbar is s/w configurable, the initial settings of the
crossbar cannot be assumed to be sane. This implies that:
a) On initialization all un-reserved crossbars must be initialized to
   a known 'safe' value.
b) When unmapping the interrupt, the safe value must be written to
   ensure that the crossbar mapping matches with interrupt controller
   usage.

So provide a ti,irqs-safe-map property to map such an safe value if
'0' is not safe(reserved) for the platform and use it during init and
unmap.

Signed-off-by: Nishanth Menon <nm@ti.com>