irqchip: crossbar allow for quirky hardware with direct hardwiring of GIC
On certain platforms such as DRA7, certain SPIs are direct wired to
hardware blocks bypassing crossbar. This is *NOT* supposed to be the
expectation of crossbar hardware usage. However, these need to be
supported for all irqs to be usable.

To allow for this, define SPI space as SPI_START to int_max (max number
of SPI interrupts) to be direct mapped and anything greater than int_max
to be crossbar numbers. Only direct wired interrupts are permissible in
this configuration.

NOTE: this requires corresponding dts changes to ensure this mapping is
adhered to.

Signed-off-by: Nishanth Menon <nm@ti.com>