io_uring: add support for polling before IO
For poll based retry, we always first attempt to read/write, or
recv/send, before arming the poll handler if we fail. If data/space is
already expected to be there, it's faster to poll first and simply let
the poll readiness drive when we attempt to do IO.

This adds IOSQE_POLL_FIRST as a flag in the sqe, which tells io_uring
to not attempt to do IO initially, but rather just arm the poll handler
and let that drive the data processing.

Signed-off-by: Jens Axboe <axboe@kernel.dk>