ARM: dts: keystone: k2g-evm: Move audio oscillator assigned clock to mcasp
Setting assigned clock with a dependency on itself for k2g_clks
creates a circular dependency. Instead, Lets model the audio clock
as a assigned-clock configuration for mcasp and set it up as the
first clock assigned-clock to be satisfied. Following the standard
convention, we use null entries to indicate entries that doesn't need
to be programmed.

References:
[1] schematics: https://www.ti.com/lit/df/sprr302/sprr302.pdf
[2] TRM: https://www.ti.com/lit/ug/spruhy8i/spruhy8i.pdf

Signed-off-by: Nishanth Menon <nm@ti.com>