TMP: arm64: dts: ti: k3-am62-sk: Enable USB and DSS root nodes as enabled for E1 board
Lets temporarily use the same patches as that from wakeup tree for
enabling HDMI and USB nodes for E1 boards. NOTE: This wont really work
on E2 and beyond boards since we flipped on bits for display and USB
ports as well.

The HDMI view has a "orange" hue due to display pins mismatch in
original E1 boards.

The gpio expander stuff is weird one.

Signed-off-by: Nishanth Menon <nm@ti.com>