vdpa/mlx5: Verify wq is a valid pointer in mlx5_vdpa_suspend
mlx5_vdpa_suspend() flushes the workqueue as part of its logic. However,
if the device has been deleted while a VM was running, the workqueue
will be destroyed first and wq will become null. After the VM is destroyed,
suspend can be called and will access a null pointer.

Fix it by verifying wq is not NULL.

Fixes: cae15c2ed8e6 ("vdpa/mlx5: Implement susupend virtqueue callback")
Signed-off-by: Eli Cohen <elic@nvidia.com>

Message-Id: <20230328071810.843361-1-elic@nvidia.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Acked-by: Eugenio Pérez Martin <eperezma@redhat.com>