HACK HACK?? work around scenario where the device is disabled by default
When the device PLL is dedicated and disabled by default upfront, TI's
DM assumes that clk_get_rate is 0 (which makes sense), however, it also
assumes that without a user (set_device_state via gen_pd), the clients
are not permitted to switch on the clock either. This creates a
conundrum for linux framework.

when PD count > 1, gen_pd waves itself off saying it has no idea of
domain dependencies and let the driver manage the domains as it chooses.

However, clk framework is independent of this and assumes that the
device will be enabled if one is attempting to set a clock rate..

Uggh.. so the mess falls on driver then? I am not going to comment on
the same :(

Signed-off-by: Nishanth Menon <nm@ti.com>